+++ Text/Inflections/Dasherize.hs
@@ -9,25 +9,24 @@
 --
 -- Conversion to dasherized phrases.
 
-{-# LANGUAGE CPP               #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module Text.Inflections.Dasherize ( dasherize ) where
+module Text.Inflections.Dasherize
+  ( dasherize )
+where
 
 import Data.Text (Text)
-import Text.Inflections.Parse.Types
+import Text.Inflections.Types
 import qualified Data.Text as T
 
-#if MIN_VERSION_base(4,8,0)
-import Prelude hiding (Word)
-#endif
-
--- | Replaces underscores in a snake_cased string with dashes (hyphens).
+-- | Produce a string with words separated by dashes (hyphens).
 --
--- >>> dasherize [ Word "foo", Acronym "bar", Word "bazz" ]
+-- >>> foo  <- SomeWord <$> mkWord "foo"
+-- >>> bar  <- SomeWord <$> mkAcronym "bar"
+-- >>> bazz <- SomeWord <$> mkWord "bazz"
+-- >>> dasherize [foo,bar,bazz]
 -- "foo-bar-bazz"
 dasherize
-  :: [Word] -- ^ Input Words to separate with dashes
-  -> Text   -- ^ The dasherized String
-dasherize = T.intercalate "-" . fmap (mapWord T.toLower)
-{-# INLINE dasherize #-}
+  :: [SomeWord] -- ^ Input words to separate with dashes
+  -> Text       -- ^ The dasherized 'Text'
+dasherize = T.intercalate "-" . fmap (unSomeWord T.toLower)