+++ Text/Inflections/Ordinal.hs
@@ -12,13 +12,25 @@
 {-# LANGUAGE OverloadedStrings #-}
 
 module Text.Inflections.Ordinal
-  ( ordinal
-  , ordinalize )
+  ( ordinalize
+  , ordinal )
 where
 
 import Data.Monoid ((<>))
 import Data.Text (Text)
 import qualified Data.Text as T
+
+-- |Turns a number into an ordinal string used to denote the position in an
+-- ordered sequence such as 1st, 2nd, 3rd, 4th.
+--
+-- >>> ordinalize 1
+-- "1st"
+-- >>> ordinalize 2
+-- "2nd"
+-- >>> ordinalize 10
+-- "10th"
+ordinalize :: (Integral a, Show a) => a -> Text
+ordinalize n = T.pack (show n) <> ordinal n
 
 -- |Returns the suffix that should be added to a number to denote the position
 -- in an ordered sequence such as 1st, 2nd, 3rd, 4th.
@@ -39,16 +51,3 @@ ordinal number
   where abs_number   = abs number
         remainder10  = abs_number `mod` 10
         remainder100 = abs_number `mod` 100
-
--- |Turns a number into an ordinal string used to denote the position in an
--- ordered sequence such as 1st, 2nd, 3rd, 4th.
---
--- >>> ordinalize 1
--- "1st"
--- >>> ordinalize 2
--- "2nd"
--- >>> ordinalize 10
--- "10th"
-ordinalize :: (Integral a, Show a) => a -> Text
-ordinalize n = T.pack (show n) <> ordinal n
-{-# INLINE ordinalize #-}